Figured out the problem
When you add a single thing to the list, head or tail, they're both set to the same value, so when you delete the head, the tail is still there. Then, because there's no head but there _is_ a tail, it adds another thing to the tail, then tries to remove the head again.
But the head is already missing and it hasn't moved down the line, it's supposed to be completely empty. I had to write a completely separate case  just to handle this one error.